requirements.txt
python-telegram-bot
python-dotenv
google-genai
google-adk
gspread
google-auth
toml
